Layer 16 in the Go tree: hybrid search, query hardening, document predicates, mutating SQL, lifecycle, throughput
Implements components 63-73 of docs/kdb-spec-layer16-qtask-gaps.md in Go.

Query semantics (69): an unknown column in WHERE, ORDER BY, GROUP BY or a function argument is
now a planning error instead of silently matching nothing or being ignored; DISTINCT is applied
by the executor; comparisons are total and never panic; the parser no longer panics on malformed
input. A schemaless namespace resolves columns dynamically, so it works as a document store.

Predicates (70): dotted paths with implicit array traversal, LIKE/ILIKE with regex escaping,
IN, BETWEEN, their NOT forms, and the ARRAY_CONTAINS family.

Mutating SQL (71): UPDATE and DELETE, plus SUM/AVG/MIN/MAX and GROUP BY.

Search (63/64/65/66/67/68): a Porter-stemming analyzer, a BM25F-lite full-text index, a vector
index with an HNSW graph over an exact oracle, reciprocal-rank and weighted fusion, MATCH /
SIMILARITY / FUSE in SQL, planner index selection, index maintenance on the commit path, and a
SEARCH wire pair with its client SDK call.

Lifecycle and throughput (72/73): documents round-trip byte-exact and a non-UUID id derives a
stable uuid8; document expiry hides at head and sweeps in the background; compound unique
constraints; per-namespace write gates and concurrent frame handling per connection.

Cross-tree parity rests on shared golden fixtures under go/testdata/golden/search, which the
Kotlin suite reads too.

+++ go/kdb/index/fusion/fusion.go
@@ -76,10 +76,13 @@ func Fuse(arms []Arm, mode Mode, limit int) []index.RankedResult {
 					hi = r.Score
 				}
 			}
+			// Normalise in float64 from the float32 inputs (and round to float32 only once, at the
+			// end): both trees do the same, so fused scores are bit-identical, which the fixture
+			// tolerance of 1e-9 needs at float32 precision.
 			for _, r := range list {
 				norm := 1.0
 				if hi > lo {
-					norm = float64(r.Score-lo) / float64(hi-lo)
+					norm = (float64(r.Score) - float64(lo)) / (float64(hi) - float64(lo))
 				}
 				bump(r.DocID, w*norm)
 			}
